jobbergate-cli/pyproject.toml
[tool.poetry]
name = "jobbergate-cli"
version = "2.2.9"
description = "Jobbergate CLI Client"
authors = ["Omnivector Solutions <info@omnivector.solutions>"]
license = "MIT"
packages = [ { include = "jobbergate_cli" } ]


[tool.poetry.dependencies]
python = "^3.8"
click = "^8.0.1"
inquirer = "^2.7.0"
PyYAML = "^5.4.1"
python-dotenv = "^0.19.0"
requests = "^2.26.0"
tabulate = "^0.8.9"
urllib3 = "^1.26.7"
sentry-sdk = "^1.4.3"
boto3 = "^1.18.64"
loguru = "^0.6.0"
python-jose = "^3.3.0"
pydantic = "^1.8.2"
yarl = "^1.7.2"
httpx = "^0.22.0"
pendulum = "^2.1.2"
py-buzz = "^3.0.0"
snick = "^1.1.0"
rich = "^11.2.0"
typer = "^0.4.0"
colorama = "^0.4.4"
plummet = {extras = ["time-machine"], version = "^1.0.0"}
auto-name-enum = "^1.0.2"
pyperclip = "^1.8.2"

[tool.poetry.dev-dependencies]
black = "^21.9b0"
pytest-cov = "^3.0.0"
pytest-random-order = "^1.0.4"
pytest-freezegun = "^0.4.2"
pytest-responsemock = "^1.0.1"
flake8 = "^3.9.2"
isort = "^5.9.3"
pytest-env = "^0.6.2"
respx = "^0.19.2"
plummet = "^1.0.0"
pyproject-flake8 = "^0.0.1-alpha.2"
mypy = "^0.931"

[tool.pytest.ini_options]
minversion = "6.0"
addopts = [
    "--random-order",
    "--cov=jobbergate_cli",
    "--cov-report=term-missing",
    "--cov-fail-under=25",
]
env = [
    "JOBBERGATE_API_ENDPOINT = https://jobbergate-api-staging.omnivector.solutions",
    "JOBBERGATE_DEBUG = false",
    "AUTH0_DOMAIN = dummy_auth_domain.com",
    "AUTH0_AUDIENCE = https://dummy_auth_audience.com",
    "AUTH0_CLIENT_ID = dummy_client_id",
    "AUTH0_CLIENT_SECRET = dummy_client_secret",
    "IDENTITY_CLAIMS_KEY = https://test-domain.test",
]
testpaths = ["jobbergate_cli/test"]

[tool.black]
line-length = 120

[tool.isort]
line_length = 120
lines_after_imports = "2"
# ignore from..import vs import while sorting a section
force_sort_within_sections = "1"
profile = "black"

[tool.flake8]
max-line-length = 120
max-complexity = 40

[tool.poetry.scripts]
jobbergate = "jobbergate_cli.main:app"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
